requirements.txt
fire==0.7.0
opencv-python==4.10.0.84
numpy==1.26.4
diffusers==0.24.0
Pillow==10.0.1
websockets==13.1
streamdiffusion[tensorrt]==0.1.1
huggingface_hub==0.24.7